separamos modelos User y Room
