+++ server/controllers/usersControllers.js
@@ -1,4 +1,4 @@
-const { Users } = require("../models/models.js");
+const { Users } = require("../models/User");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
