--- server/middlewares/passport.js
@@ -1,24 +1,0 @@
-const GoogleStrategy = require("passport-google-oauth20").Strategy;
-const passport = require('passport');
-
-passport.use(
-    new GoogleStrategy(
-        {
-            clientID: process.env.CLIENT_ID,
-            clientSecret: process.env.CLIENT_SECRET,
-            callbackURL: "/auth/google/callback",
-            scope:["profile", "email"]
-        },
-        function(accesToken, refreshToken, profile, callback){
-            callback(null, profile);
-        }
-    )
-);
-
-passport.serializeUser((user, done) => {
-    done(null, user);
-});
-
-passport.deserializeUser((user, done) => {
-    done(null, user);
-});
